fix: build overrides for headless push with new env init scneario
